MIPS: JZ4740: rename irq.h to non-reserved name
Patch to solve build error when building with an output outside the
tree.
