+++ arch/mips/jz4740/irq-jz4740.h
